+++ backend/internal/example-custom-agent/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @cat-factory/example-custom-agent
 
+## 0.0.34
+
+### Patch Changes
+
+- Updated dependencies [4dd6e97]
+  - @cat-factory/agents@0.16.1
+
+## 0.0.33
+
+### Patch Changes
+
+- Updated dependencies [ea59e91]
+  - @cat-factory/kernel@0.33.0
+  - @cat-factory/agents@0.16.0
+
 ## 0.0.32
 
 ### Patch Changes